split test cases between flask-restx and controller
