+++ tests/test_sqlalchemy_controller_date.py
@@ -1,8 +1,5 @@
 import datetime
-import collections.abc
 
-import flask
-import flask_restplus
 import pytest
 import sqlalchemy
 
@@ -10,7 +7,7 @@ import layabase
 
 
 @pytest.fixture
-def controller():
+def controller() -> layabase.CRUDController:
     class TestTable:
         __tablename__ = "test"
 
@@ -23,353 +20,7 @@ def controller():
     return controller
 
 
-@pytest.fixture
-def app(controller):
-    application = flask.Flask(__name__)
-    application.testing = True
-    api = flask_restplus.Api(application)
-    namespace = api.namespace("Test", path="/")
-
-    controller.namespace(namespace)
-
-    @namespace.route("/test")
-    class TestResource(flask_restplus.Resource):
-        @namespace.expect(controller.query_get_parser)
-        @namespace.marshal_with(controller.get_response_model)
-        def get(self):
-            return []
-
-        @namespace.expect(controller.json_post_model)
-        def post(self):
-            return []
-
-        @namespace.expect(controller.json_put_model)
-        def put(self):
-            return []
-
-        @namespace.expect(controller.query_delete_parser)
-        def delete(self):
-            return []
-
-    @namespace.route("/test_parsers")
-    class TestParsersResource(flask_restplus.Resource):
-        @namespace.expect(controller.query_get_parser)
-        def get(self):
-            return {
-                field: [str(value) for value in values]
-                if isinstance(values, collections.abc.Iterable)
-                else values
-                for field, values in controller.query_get_parser.parse_args().items()
-            }
-
-        @namespace.expect(controller.query_delete_parser)
-        def delete(self):
-            return {
-                field: [str(value) for value in values]
-                for field, values in controller.query_delete_parser.parse_args().items()
-            }
-
-    return application
-
-
-def test_open_api_definition(client):
-    response = client.get("/swagger.json")
-    assert response.json == {
-        "swagger": "2.0",
-        "basePath": "/",
-        "paths": {
-            "/test": {
-                "post": {
-                    "responses": {"200": {"description": "Success"}},
-                    "operationId": "post_test_resource",
-                    "parameters": [
-                        {
-                            "name": "payload",
-                            "required": True,
-                            "in": "body",
-                            "schema": {
-                                "$ref": "#/definitions/TestTable_PostRequestModel"
-                            },
-                        }
-                    ],
-                    "tags": ["Test"],
-                },
-                "put": {
-                    "responses": {"200": {"description": "Success"}},
-                    "operationId": "put_test_resource",
-                    "parameters": [
-                        {
-                            "name": "payload",
-                            "required": True,
-                            "in": "body",
-                            "schema": {
-                                "$ref": "#/definitions/TestTable_PutRequestModel"
-                            },
-                        }
-                    ],
-                    "tags": ["Test"],
-                },
-                "delete": {
-                    "responses": {"200": {"description": "Success"}},
-                    "operationId": "delete_test_resource",
-                    "parameters": [
-                        {
-                            "name": "key",
-                            "in": "query",
-                            "type": "array",
-                            "items": {"type": "string"},
-                            "collectionFormat": "multi",
-                        },
-                        {
-                            "name": "date_str",
-                            "in": "query",
-                            "type": "array",
-                            "format": "date",
-                            "items": {"type": "string"},
-                            "collectionFormat": "multi",
-                        },
-                        {
-                            "name": "datetime_str",
-                            "in": "query",
-                            "type": "array",
-                            "format": "date-time",
-                            "items": {"type": "string"},
-                            "collectionFormat": "multi",
-                        },
-                    ],
-                    "tags": ["Test"],
-                },
-                "get": {
-                    "responses": {
-                        "200": {
-                            "description": "Success",
-                            "schema": {
-                                "$ref": "#/definitions/TestTable_GetResponseModel"
-                            },
-                        }
-                    },
-                    "operationId": "get_test_resource",
-                    "parameters": [
-                        {
-                            "name": "key",
-                            "in": "query",
-                            "type": "array",
-                            "items": {"type": "string"},
-                            "collectionFormat": "multi",
-                        },
-                        {
-                            "name": "date_str",
-                            "in": "query",
-                            "type": "array",
-                            "format": "date",
-                            "items": {"type": "string"},
-                            "collectionFormat": "multi",
-                        },
-                        {
-                            "name": "datetime_str",
-                            "in": "query",
-                            "type": "array",
-                            "format": "date-time",
-                            "items": {"type": "string"},
-                            "collectionFormat": "multi",
-                        },
-                        {
-                            "name": "limit",
-                            "in": "query",
-                            "type": "integer",
-                            "minimum": 0,
-                            "exclusiveMinimum": True,
-                        },
-                        {
-                            "name": "offset",
-                            "in": "query",
-                            "type": "integer",
-                            "minimum": 0,
-                        },
-                        {
-                            "name": "order_by",
-                            "in": "query",
-                            "type": "array",
-                            "items": {"type": "string"},
-                            "collectionFormat": "multi",
-                        },
-                        {
-                            "name": "X-Fields",
-                            "in": "header",
-                            "type": "string",
-                            "format": "mask",
-                            "description": "An optional fields mask",
-                        },
-                    ],
-                    "tags": ["Test"],
-                },
-            },
-            "/test_parsers": {
-                "delete": {
-                    "responses": {"200": {"description": "Success"}},
-                    "operationId": "delete_test_parsers_resource",
-                    "parameters": [
-                        {
-                            "name": "key",
-                            "in": "query",
-                            "type": "array",
-                            "items": {"type": "string"},
-                            "collectionFormat": "multi",
-                        },
-                        {
-                            "name": "date_str",
-                            "in": "query",
-                            "type": "array",
-                            "format": "date",
-                            "items": {"type": "string"},
-                            "collectionFormat": "multi",
-                        },
-                        {
-                            "name": "datetime_str",
-                            "in": "query",
-                            "type": "array",
-                            "format": "date-time",
-                            "items": {"type": "string"},
-                            "collectionFormat": "multi",
-                        },
-                    ],
-                    "tags": ["Test"],
-                },
-                "get": {
-                    "responses": {"200": {"description": "Success"}},
-                    "operationId": "get_test_parsers_resource",
-                    "parameters": [
-                        {
-                            "name": "key",
-                            "in": "query",
-                            "type": "array",
-                            "items": {"type": "string"},
-                            "collectionFormat": "multi",
-                        },
-                        {
-                            "name": "date_str",
-                            "in": "query",
-                            "type": "array",
-                            "format": "date",
-                            "items": {"type": "string"},
-                            "collectionFormat": "multi",
-                        },
-                        {
-                            "name": "datetime_str",
-                            "in": "query",
-                            "type": "array",
-                            "format": "date-time",
-                            "items": {"type": "string"},
-                            "collectionFormat": "multi",
-                        },
-                        {
-                            "name": "limit",
-                            "in": "query",
-                            "type": "integer",
-                            "minimum": 0,
-                            "exclusiveMinimum": True,
-                        },
-                        {
-                            "name": "offset",
-                            "in": "query",
-                            "type": "integer",
-                            "minimum": 0,
-                        },
-                        {
-                            "name": "order_by",
-                            "in": "query",
-                            "type": "array",
-                            "items": {"type": "string"},
-                            "collectionFormat": "multi",
-                        },
-                    ],
-                    "tags": ["Test"],
-                },
-            },
-        },
-        "info": {"title": "API", "version": "1.0"},
-        "produces": ["application/json"],
-        "consumes": ["application/json"],
-        "tags": [{"name": "Test"}],
-        "definitions": {
-            "TestTable_PostRequestModel": {
-                "required": ["key"],
-                "properties": {
-                    "key": {
-                        "type": "string",
-                        "readOnly": False,
-                        "example": "sample_value",
-                    },
-                    "date_str": {
-                        "type": "string",
-                        "format": "date",
-                        "readOnly": False,
-                        "example": "2017-09-24",
-                    },
-                    "datetime_str": {
-                        "type": "string",
-                        "format": "date-time",
-                        "readOnly": False,
-                        "example": "2017-09-24T15:36:09",
-                    },
-                },
-                "type": "object",
-            },
-            "TestTable_PutRequestModel": {
-                "required": ["key"],
-                "properties": {
-                    "key": {
-                        "type": "string",
-                        "readOnly": False,
-                        "example": "sample_value",
-                    },
-                    "date_str": {
-                        "type": "string",
-                        "format": "date",
-                        "readOnly": False,
-                        "example": "2017-09-24",
-                    },
-                    "datetime_str": {
-                        "type": "string",
-                        "format": "date-time",
-                        "readOnly": False,
-                        "example": "2017-09-24T15:36:09",
-                    },
-                },
-                "type": "object",
-            },
-            "TestTable_GetResponseModel": {
-                "required": ["key"],
-                "properties": {
-                    "key": {
-                        "type": "string",
-                        "readOnly": False,
-                        "example": "sample_value",
-                    },
-                    "date_str": {
-                        "type": "string",
-                        "format": "date",
-                        "readOnly": False,
-                        "example": "2017-09-24",
-                    },
-                    "datetime_str": {
-                        "type": "string",
-                        "format": "date-time",
-                        "readOnly": False,
-                        "example": "2017-09-24T15:36:09",
-                    },
-                },
-                "type": "object",
-            },
-        },
-        "responses": {
-            "ParseError": {"description": "When a mask can't be parsed"},
-            "MaskError": {"description": "When any error occurs on mask"},
-        },
-    }
-
-
-def test_put_is_updating_date(controller):
+def test_put_is_updating_date(controller: layabase.CRUDController):
     controller.post(
         {
             "key": "my_key1",
@@ -404,7 +55,7 @@ def test_put_is_updating_date(controller):
     ] == controller.get({"date_str": "2018-06-01"})
 
 
-def test_get_date_is_handled_for_valid_date(controller):
+def test_get_date_is_handled_for_valid_date(controller: layabase.CRUDController):
     controller.post(
         {
             "key": "my_key1",
@@ -421,7 +72,7 @@ def test_get_date_is_handled_for_valid_date(controller):
     ] == controller.get({"date_str": datetime.date(2017, 5, 15)})
 
 
-def test_get_date_is_handled_for_unused_date(controller):
+def test_get_date_is_handled_for_unused_date(controller: layabase.CRUDController):
     controller.post(
         {
             "key": "my_key1",
@@ -433,7 +84,7 @@ def test_get_date_is_handled_for_unused_date(controller):
     assert [] == controller.get({"date_str": d})
 
 
-def test_get_date_is_handled_for_valid_datetime(controller):
+def test_get_date_is_handled_for_valid_datetime(controller: layabase.CRUDController):
     controller.post(
         {
             "key": "my_key1",
@@ -450,7 +101,7 @@ def test_get_date_is_handled_for_valid_datetime(controller):
     ] == controller.get({"datetime_str": datetime.datetime(2016, 9, 23, 23, 59, 59)})
 
 
-def test_get_date_is_handled_for_unused_datetime(controller):
+def test_get_date_is_handled_for_unused_datetime(controller: layabase.CRUDController):
     controller.post(
         {
             "key": "my_key1",
@@ -460,28 +111,3 @@ def test_get_date_is_handled_for_unused_datetime(controller):
     )
     dt = datetime.datetime.strptime("2016-09-24T23:59:59", "%Y-%m-%dT%H:%M:%S")
     assert [] == controller.get({"datetime_str": dt})
-
-
-def test_query_get_parser(client):
-    response = client.get(
-        "/test_parsers?key=12&date_str=2017-05-15&datetime_str=2016-09-23T23:59:59&limit=1&order_by=key&offset=0"
-    )
-    assert response.json == {
-        "date_str": ["2017-05-15"],
-        "datetime_str": ["2016-09-23 23:59:59"],
-        "key": ["12"],
-        "limit": 1,
-        "offset": 0,
-        "order_by": ["key"],
-    }
-
-
-def test_query_delete_parser(client):
-    response = client.delete(
-        "/test_parsers?key=12&date_str=2017-05-15&datetime_str=2016-09-23T23:59:59&"
-    )
-    assert response.json == {
-        "date_str": ["2017-05-15"],
-        "datetime_str": ["2016-09-23 23:59:59"],
-        "key": ["12"],
-    }
